test(gateway): run Codex boundary proof in E2E
